Add documentation into network code
