+++ network/Cargo.toml
@@ -35,3 +35,6 @@ snarkos-testing = { path = "../testing" }
 
 serial_test = { version = "0.4.0" }
 tokio-test = { version = "0.2.0" }
+
+[build-dependencies]
+rustc_version = "0.2"
