+++ network/build.rs
@@ -14,44 +14,15 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkOS library. If not, see <https://www.gnu.org/licenses/>.
 
-pub mod block;
-pub use block::*;
+// Detect the rustc channel
+use rustc_version::{version_meta, Channel};
 
-pub mod getblock;
-pub use getblock::*;
-
-pub mod getmemorypool;
-pub use getmemorypool::*;
-
-pub mod getpeers;
-pub use getpeers::*;
-
-pub mod getsync;
-pub use getsync::*;
-
-pub mod memorypool;
-pub use memorypool::*;
-
-pub mod peers;
-pub use peers::*;
-
-pub mod ping;
-pub use ping::*;
-
-pub mod pong;
-pub use pong::*;
-
-pub mod sync;
-pub use sync::*;
-
-pub mod syncblock;
-pub use syncblock::*;
-
-pub mod transaction;
-pub use transaction::*;
-
-pub mod verack;
-pub use verack::*;
-
-pub mod version;
-pub use version::*;
+fn main() {
+    // Set cfg flags depending on release channel
+    match version_meta().unwrap().channel {
+        Channel::Stable => println!("cargo:rustc-cfg=stable"),
+        Channel::Beta => println!("cargo:rustc-cfg=beta"),
+        Channel::Nightly => println!("cargo:rustc-cfg=nightly"),
+        Channel::Dev => println!("cargo:rustc-cfg=rustc_dev"),
+    }
+}
