+++ network/documentation/network_messages/block.md
@@ -1,5 +1,9 @@
 Send a block to a peer.
 
+### Message Name
+
+`block`
+
 ### Payload
 
 The serialized bytes of the block.
