+++ network/documentation/network_messages/get_block.md
@@ -1,5 +1,9 @@
 A request for a block with the specified hash.
 
+### Message Name
+
+`getblock`
+
 ### Payload
 
 |   Parameter  | Type  |              Description              |
