+++ network/documentation/network_messages/get_memory_pool.md
@@ -1,5 +1,9 @@
 A request for a peer's memory pool transactions.
 
+### Message Name
+
+`getmemorypool`
+
 ### Payload
 
 `None`
