+++ network/documentation/network_messages/get_peers.md
@@ -1,5 +1,9 @@
 A request for a list of the peer's connected peer addresses.
 
+### Message Name
+
+`getpeers`
+
 ### Payload
 
 `None`
