+++ network/documentation/network_messages/pong.md
@@ -1,4 +1,8 @@
-A response to a Ping request.
+A response to a `Ping` request.
+
+### Message Name
+
+`pong`
 
 ### Payload
