+++ network/documentation/network_messages/sync_block.md
@@ -1,4 +1,8 @@
-A response to a GetBlock request.
+A response to a `GetBlock` request.
+
+### Message Name
+
+`syncblock`
 
 ### Payload
