+++ network/documentation/network_messages/verack.md
@@ -1,4 +1,8 @@
-A handshake response to a Version message.
+A handshake response to a `Version` message.
+
+### Message Name
+
+`verack`
 
 ### Payload
