+++ network/src/message_types/block.rs
@@ -16,50 +16,47 @@
 
 use crate::message::{Message, MessageName};
 use snarkos_errors::network::message::MessageError;
-use snarkos_objects::BlockHeaderHash;
 
-/// A response to a GetSync message.
-/// See network/protocol/sync.rs for more details.
+#[cfg_attr(nightly, doc(include = "../../documentation/network_messages/block.md"))]
 #[derive(Debug, PartialEq, Clone)]
-pub struct Sync {
-    /// Known hashes of blocks to share
-    pub block_hashes: Vec<BlockHeaderHash>,
+pub struct Block {
+    /// Serialized block data
+    pub data: Vec<u8>,
 }
 
-impl Sync {
-    pub fn new(block_hashes: Vec<BlockHeaderHash>) -> Self {
-        Self { block_hashes }
+impl Block {
+    pub fn new(data: Vec<u8>) -> Self {
+        Self { data }
     }
 }
 
-impl Message for Sync {
+impl Message for Block {
     fn name() -> MessageName {
-        MessageName::from("sync")
+        MessageName::from("block")
     }
 
     fn deserialize(vec: Vec<u8>) -> Result<Self, MessageError> {
         Ok(Self {
-            block_hashes: bincode::deserialize(&vec)?,
+            data: bincode::deserialize(&vec)?,
         })
     }
 
     fn serialize(&self) -> Result<Vec<u8>, MessageError> {
-        Ok(bincode::serialize(&self.block_hashes)?)
+        Ok(bincode::serialize(&self.data)?)
     }
 }
 
 #[cfg(test)]
 mod tests {
     use super::*;
-    use snarkos_testing::consensus::BLOCK_1_HEADER_HASH;
+    use snarkos_testing::consensus::BLOCK_1;
 
     #[test]
-    fn test_sync() {
-        let data = BlockHeaderHash::new(BLOCK_1_HEADER_HASH.to_vec());
-        let message = Sync::new(vec![data]);
+    fn test_block() {
+        let message = Block::new(BLOCK_1.to_vec());
 
         let serialized = message.serialize().unwrap();
-        let deserialized = Sync::deserialize(serialized).unwrap();
+        let deserialized = Block::deserialize(serialized).unwrap();
 
         assert_eq!(message, deserialized);
     }
