+++ network/src/message_types/getblock.rs
@@ -16,51 +16,49 @@
 
 use crate::message::{Message, MessageName};
 use snarkos_errors::network::message::MessageError;
+use snarkos_objects::BlockHeaderHash;
 
-use chrono::{DateTime, Utc};
-use std::{collections::HashMap, net::SocketAddr};
-
-/// A response to a GetPeers request.
+#[cfg_attr(nightly, doc(include = "../../documentation/network_messages/get_block.md"))]
 #[derive(Debug, PartialEq, Clone)]
-pub struct Peers {
-    /// A list of gossiped peer addresses and their last seen dates
-    pub addresses: HashMap<SocketAddr, DateTime<Utc>>,
+pub struct GetBlock {
+    /// Header hash of requested block
+    pub block_hash: BlockHeaderHash,
 }
 
-impl Peers {
-    pub fn new(addresses: HashMap<SocketAddr, DateTime<Utc>>) -> Self {
-        Self { addresses }
+impl GetBlock {
+    pub fn new(block_hash: BlockHeaderHash) -> Self {
+        Self { block_hash }
     }
 }
 
-impl Message for Peers {
+impl Message for GetBlock {
     fn name() -> MessageName {
-        MessageName::from("peers")
+        MessageName::from("getblock")
     }
 
     fn deserialize(vec: Vec<u8>) -> Result<Self, MessageError> {
         Ok(Self {
-            addresses: bincode::deserialize(&vec)?,
+            block_hash: bincode::deserialize(&vec)?,
         })
     }
 
     fn serialize(&self) -> Result<Vec<u8>, MessageError> {
-        Ok(bincode::serialize(&self.addresses)?)
+        Ok(bincode::serialize(&self.block_hash)?)
     }
 }
 
 #[cfg(test)]
 mod tests {
     use super::*;
+    use snarkos_testing::consensus::BLOCK_1_HEADER_HASH;
 
     #[test]
-    fn test_peers() {
-        let message = Peers {
-            addresses: HashMap::<SocketAddr, DateTime<Utc>>::new(),
-        };
+    fn test_block() {
+        let block_hash = BlockHeaderHash::new(BLOCK_1_HEADER_HASH.to_vec());
+        let message = GetBlock::new(block_hash);
 
         let serialized = message.serialize().unwrap();
-        let deserialized = Peers::deserialize(serialized).unwrap();
+        let deserialized = GetBlock::deserialize(serialized).unwrap();
 
         assert_eq!(message, deserialized);
     }
