+++ network/src/message_types/peers.rs
@@ -17,49 +17,35 @@
 use crate::message::{Message, MessageName};
 use snarkos_errors::network::message::MessageError;
 
-use byteorder::{BigEndian, ReadBytesExt, WriteBytesExt};
-use rand::Rng;
-use std::io::Cursor;
+use chrono::{DateTime, Utc};
+use std::{collections::HashMap, net::SocketAddr};
 
-/// A ping protocol request for a Pong.
-/// See network/protocol/ping_protocol.rs for more details.
+#[cfg_attr(nightly, doc(include = "../../documentation/network_messages/peers.md"))]
 #[derive(Debug, PartialEq, Clone)]
-pub struct Ping {
-    /// Unique ping protocol identifier
-    pub nonce: u64,
+pub struct Peers {
+    /// A list of gossiped peer addresses and their last seen dates
+    pub addresses: HashMap<SocketAddr, DateTime<Utc>>,
 }
 
-impl Ping {
-    pub fn new() -> Self {
-        let mut rng = rand::thread_rng();
-        Self {
-            nonce: rng.gen::<u64>(),
-        }
+impl Peers {
+    pub fn new(addresses: HashMap<SocketAddr, DateTime<Utc>>) -> Self {
+        Self { addresses }
     }
 }
 
-impl Message for Ping {
+impl Message for Peers {
     fn name() -> MessageName {
-        MessageName::from("ping")
+        MessageName::from("peers")
     }
 
     fn deserialize(vec: Vec<u8>) -> Result<Self, MessageError> {
-        if vec.len() != 8 {
-            return Err(MessageError::InvalidLength(vec.len(), 8));
-        }
-
-        let mut reader = Cursor::new(vec);
-
         Ok(Self {
-            nonce: reader.read_u64::<BigEndian>().expect("unable to read u64"),
+            addresses: bincode::deserialize(&vec)?,
         })
     }
 
     fn serialize(&self) -> Result<Vec<u8>, MessageError> {
-        let mut writer = vec![];
-        writer.write_u64::<BigEndian>(self.nonce)?;
-
-        Ok(writer)
+        Ok(bincode::serialize(&self.addresses)?)
     }
 }
 
@@ -68,11 +54,13 @@ mod tests {
     use super::*;
 
     #[test]
-    fn test_ping() {
-        let message = Ping::new();
+    fn test_peers() {
+        let message = Peers {
+            addresses: HashMap::<SocketAddr, DateTime<Utc>>::new(),
+        };
 
         let serialized = message.serialize().unwrap();
-        let deserialized = Ping::deserialize(serialized).unwrap();
+        let deserialized = Peers::deserialize(serialized).unwrap();
 
         assert_eq!(message, deserialized);
     }
