+++ network/src/message_types/sync.rs
@@ -16,47 +16,49 @@
 
 use crate::message::{Message, MessageName};
 use snarkos_errors::network::message::MessageError;
+use snarkos_objects::BlockHeaderHash;
 
-/// A response to a GetMemoryPool request.
+#[cfg_attr(nightly, doc(include = "../../documentation/network_messages/sync.md"))]
 #[derive(Debug, PartialEq, Clone)]
-pub struct MemoryPool {
-    /// Vector of memory pool transactions
-    pub transactions: Vec<Vec<u8>>,
+pub struct Sync {
+    /// Known hashes of blocks to share
+    pub block_hashes: Vec<BlockHeaderHash>,
 }
 
-impl MemoryPool {
-    pub fn new(transactions: Vec<Vec<u8>>) -> Self {
-        Self { transactions }
+impl Sync {
+    pub fn new(block_hashes: Vec<BlockHeaderHash>) -> Self {
+        Self { block_hashes }
     }
 }
 
-impl Message for MemoryPool {
+impl Message for Sync {
     fn name() -> MessageName {
-        MessageName::from("memorypool")
+        MessageName::from("sync")
     }
 
     fn deserialize(vec: Vec<u8>) -> Result<Self, MessageError> {
         Ok(Self {
-            transactions: bincode::deserialize(&vec)?,
+            block_hashes: bincode::deserialize(&vec)?,
         })
     }
 
     fn serialize(&self) -> Result<Vec<u8>, MessageError> {
-        Ok(bincode::serialize(&self.transactions)?)
+        Ok(bincode::serialize(&self.block_hashes)?)
     }
 }
 
 #[cfg(test)]
 mod tests {
     use super::*;
-    use snarkos_testing::consensus::TRANSACTION_1;
+    use snarkos_testing::consensus::BLOCK_1_HEADER_HASH;
 
     #[test]
-    fn test_memory_pool() {
-        let message = MemoryPool::new(vec![TRANSACTION_1.to_vec()]);
+    fn test_sync() {
+        let data = BlockHeaderHash::new(BLOCK_1_HEADER_HASH.to_vec());
+        let message = Sync::new(vec![data]);
 
         let serialized = message.serialize().unwrap();
-        let deserialized = MemoryPool::deserialize(serialized).unwrap();
+        let deserialized = Sync::deserialize(serialized).unwrap();
 
         assert_eq!(message, deserialized);
     }
