+++ network/src/message_types/version.rs
@@ -21,8 +21,7 @@ use chrono::Utc;
 use rand::Rng;
 use std::net::SocketAddr;
 
-/// A handshake request for a Verack.
-/// See network/protocol/handshake.rs for more details.
+#[cfg_attr(nightly, doc(include = "../../documentation/network_messages/version.md"))]
 #[derive(Debug, PartialEq, Clone)]
 pub struct Version {
     /// The network version number
