+++ network/src/protocol/handshake.rs
@@ -14,9 +14,9 @@
 // You should have received a copy of the GNU General Public License
 // along with the snarkOS library. If not, see <https://www.gnu.org/licenses/>.
 
-use crate::message::{
-    types::{Verack, Version},
-    Channel,
+use crate::{
+    message::Channel,
+    message_types::{Verack, Version},
 };
 use snarkos_errors::network::HandshakeError;
